+++ src/core/task/Task.ts
@@ -516,6 +516,7 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 		this.enableCheckpoints = enableCheckpoints
 		this.checkpointTimeout = checkpointTimeout
 
+		this.rootTask = rootTask
 		this.parentTask = parentTask
 		this.taskNumber = taskNumber
 		this.initialStatus = initialStatus
@@ -2568,6 +2569,7 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 				JSON.stringify({
 					apiProtocol,
 					profileName: this.taskApiConfigName,
+					modelId,
 				}),
 			)
 
@@ -2648,6 +2650,7 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 			this.clineMessages[lastApiReqIndex].text = JSON.stringify({
 				apiProtocol,
 				profileName: this.taskApiConfigName,
+				modelId,
 			} satisfies ClineApiReqInfo)
 
 			await this.saveClineMessages()